+++ packages/query/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Custom OrderBy plugin
+- Custom OrderBy plugin to improve dictionary query performance (#1907)
 
 ## [2.3.0] - 2023-07-04
 ### Changed