+++ packages/query/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Custom OrderBy plugin to improve dictionary query performance (#1907)
 
 ## [2.3.0] - 2023-07-04
 ### Changed