+++ packages/query/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.4.0] - 2023-07-31
+### Fixed
+- Avoid get `metadataTableName` call system table (#1893)
+- Update license (#1891)
 ### Added
 - Custom OrderBy plugin to improve dictionary query performance (#1907)
 
@@ -252,7 +257,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - overwrite plugin to fix one to one unique key check
 - update query publish and docker build process
 
-[Unreleased]: https://github.com/subquery/subql/compare/query/2.3.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/query/2.4.0...HEAD
+[2.4.0]: https://github.com/subquery/subql/compare/query/2.3.0...query/2.4.0
 [2.3.0]: https://github.com/subquery/subql/compare/query/2.2.0...query/2.3.0
 [2.2.0]: https://github.com/subquery/subql/compare/query/2.1.0...query/2.2.0
 [2.1.0]: https://github.com/subquery/subql/compare/query/2.0.1..query/2.1.0